recommendation/requirements.txt
flask==2.3.3
flask-cors==4.0.0
python-dotenv==1.0.0
scikit-learn==1.3.2
pandas==2.1.4
numpy==1.26.4
gunicorn==21.2.0